+++ src/consumer/bit-json/consumer-bit-json.js
@@ -21,6 +21,7 @@ export default class ConsumerBitJson extends AbstractBitJson {
   compiler: string;
   tester: string;
   dependencies: {[string]: string};
+  lang: string;
 
   write({ bitDir, override = true }: { bitDir: string, override?: boolean }): Promise<boolean> {
     return new Promise((resolve, reject) => {
@@ -54,13 +55,14 @@ export default class ConsumerBitJson extends AbstractBitJson {
   }
 
   static fromPlainObject(object: Object) {
-    const { sources, env, dependencies } = object;
+    const { sources, env, dependencies, lang } = object;
     return new ConsumerBitJson({
       impl: R.prop('impl', sources),
       spec: R.prop('spec', sources),
       miscFiles: R.prop('misc', sources),
       compiler: R.prop('compiler', env),
       tester: R.prop('tester', env),
+      lang,
       dependencies,
     });
   }